+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useJira';
 export * from './useJiraDashboard';
 export * from './useJiraUpdatesPolling';
+export * from './usePageAutoRefresh';